+++ src/main/java/by/epam/naumovich/film_ordering/service/IFilmService.java
@@ -152,5 +152,5 @@ public interface IFilmService {
 	 * 
 	 * @return number of pages
 	 */
-	int countPages();
+	long countPages();
 }
