+++ src/main/java/by/epam/naumovich/film_ordering/service/IUserService.java
@@ -144,5 +144,5 @@ public interface IUserService {
 	 * 
 	 * @return number of pages
 	 */
-	int countPages();
+	long countPages();
 }
